Refactor: extract complex onboarding components
